+++ protoc_plugin/lib/src/enum_generator.dart
@@ -103,6 +103,10 @@ class EnumGenerator extends ProtobufContainer {
   static const int _enumValueTag = 2;
 
   void generate(IndentingWriter out) {
+    final comment = fileGen?.commentBlock(fieldPath!);
+    if (comment != null) {
+      out.println(comment);
+    }
     out.addAnnotatedBlock(
         'class $classname extends $protobufImportPrefix.ProtobufEnum {',
         '}\n', [